requirements.txt
aiohttp==3.7.4.post0
astroid==2.5.1
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.6
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==3.4.6
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.2
dicttoxml==1.7.4
docutils==0.16
future==0.18.2
geocoder==1.38.1
geoip2==4.1.0
idna==2.10
importlib-metadata==3.7.3
IP2Location==8.5.1
isort==5.8.0
jeepney==0.6.0
keyring==23.0.0
kiwisolver==1.3.1
lazy-object-proxy==1.5.2
lxml==4.6.3
matplotlib==3.4.1
maxminddb==2.0.3
mccabe==0.6.1
multidict==5.1.0
numpy==1.20.2
packaging==20.9
pandas==1.2.4
Pillow==8.2.0
pip-review==1.1.0
pkg-resources==0.0.0
pkginfo==1.7.0
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.8.1
pylint==2.7.2
pyparsing==2.4.7
pyquery==1.4.3
python-dateutil==2.8.1
pytz==2021.1
ratelim==0.1.6
readme-renderer==29.0
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.4.0
ripe.atlas.cousteau==1.4.2
SecretStorage==3.3.1
selenium==3.141.0
six==1.15.0
socketIO-client==0.7.2
toml==0.10.2
tqdm==4.59.0
twine==3.4.1
typed-ast==1.4.2
typing==3.7.4.3
typing-extensions==3.7.4.3
urllib3==1.26.4
webencodings==0.5.1
websocket-client==0.58.0
wrapt==1.12.1
yarl==1.6.3
zipp==3.4.1
